+++ 3o14/src/models/follow.ts
@@ -1,0 +1,5 @@
+export interface Follow {
+  following_id: number;
+  follower_id: number;
+  created: string;
+}
