+++ 3o14/src/models/post.ts
@@ -1,0 +1,8 @@
+export interface Post {
+  id: number;
+  uri: string;
+  actor_id: number;
+  content: string;
+  url: string | null;
+  created: string;
+}
