+++ 3o14/src/models/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+  id: number;
+  username: string;
+}
